SCHEDULING-1590 refactor Scheduler Database
Database support for storing scheduler data was rewritten.

Instead of directly mapping internal classes containing logic to the
database tables special classes intended just for database storage were
added (there are put in the package
'org.ow2.proactive.scheduler.core.db'). Conversion to/from internal
classes requires some more coding but this approach allows to implement
optimal schema. As result new schema has 12 tables instead of > 60
before and it works with Derby, MySQL, Oracle and H2 databases (didn't
test with others).
All operations with database are executed using new class
SchedulerDBManager, as opposed to the old DatabaseManager it contains
fine-grained methods (e.g. jobTaskStarted, changeJobPriority, etc)
instead of general method 'update', 'register'.  In this way some
operations can be implemented much more efficiently (for some operations
instead of updating all object graph just few attributes are updated),
client of SchedulerDBManager don't need to know details of DB schema and
this allowed to get rid of native SQL queries. Set of tests for
SchedulerDBManager and related classes was added, they test all DB
operations without starting scheduler.  

Some others notable changes:
- latest versions of Hibernate and Derby are used. Note: after migration
it is necessary to use another values for derby and MySql Hibernate
dialects: org.hibernate.dialect.DerbyTenFiveDialect and
org.hibernate.dialect.MySQL5InnoDBDialect.
- scheduler supports mode without using persistent storage, before there
was special class SchedulerEmptyDatabaseManager and it didn't support
all operations. Now this mode is implemented using regular
SchedulerDBManager with configuration using H2 in-memory database. 
- old classes related to scheduler DB support were removed
(SchedulerDatabaseManager, etc), hibernate annotations were removed from
scheduler classes
- logic of SchedulerCore.recover was moved to the special class
SchedulerStateRecoverHelper, this allows to test recovering code without
instantiating scheduler
- there is no global static SchedulerDBManager instance,
SchedulerDBManager is created by SchedulerFrontend and passed to the
classes using it
- JobResult isn't stored, it is loaded when needed

+++ src/scheduler/tests/functionaltests/TestExportVars.java
@@ -104,7 +104,7 @@ public class TestExportVars extends FunctionalTest {
         taskid = "task5";
         r = res.getResult(taskid);
         // exception in post script evaluation
-        Assert.assertEquals(0, (Integer) r.value());
+        Assert.assertEquals((Integer) 0, (Integer) r.value());
 
     }
